Refactor WYSIWYG editor structure by consolidating grapesBlocks functionality into gjs-blocks, updating component registry and editor settings, and enhancing utility functions for improved performance and maintainability.
