+++ .changeset/odd-roses-fix.md
@@ -1,5 +1,6 @@
 ---
 "@studiocms/html": patch
+"@studiocms/wysiwyg": patch
 ---
 
 Optimize editor imports to use dynamic imports to help tree shaking
