+++ packages/@studiocms/wysiwyg/src/common/gjs-editor-utils.ts
